making view, model class
